requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
asyncpg==0.28.0
awslogs==0.14.0
bcrypt==4.0.1
blinker==1.7.0
boto3==1.34.49
botocore==1.34.49
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
databases==0.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.0
fastapi-mail==1.4.1
greenlet==3.0.0
h11==0.14.0
httptools==0.6.1
idna==3.4
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.3
numpy==1.26.2
pandas==2.1.3
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
termcolor==2.4.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.23.2
watchfiles==0.21.0
websockets==11.0.3
wheel==0.43.0